添加login.py和修改zhuti1.0.py
