+++ .idea/.name
@@ -1,0 +1,1 @@
+zhuti1.0.py
